Harder than it should have been
Hard because I didn't account for the fact that moving around the node
mean that I first had to splice it out, then do all calculations on a
list that had one fewer element. This was one where I used Norvig to
check that my input was right because I was starting to suspect it was
wrong (it wasn't). He too was bitten by a list that was too long by
1. My circular list implementation was not his, but the problem was
identical.